generate ASTs for coupled formulae
Reimplement the reduction functions as an AST-to-AST operation instead
of directly generating LaTeX code. Makes the result more usable and
allows post-processing, e.g., auto-generating code from the coupled
formulae.

This is somewhat work-in-progress. LaTeX output works, but some errors
in the results are to be expected. Also, the reimplementation currently
lacks the permutation functionality that exploits antisymmetry of the
tensors to simplify the coupled expressions. However, it is now possible
to define custom coupling schemes, such as cross-coupled tensors.