add tensor coupling for nonscalar tensors
Nonscalar tensors retain an extra rank index that leaves the Yutsis
graph open, so the usual reduction is impossible. We have to
additionally form irreducible tensor products, coupling the ranks to the
left-hand side rank, in order to close the graph. Currently, we do this
by simply coupling all tensors left to right to intermediate ranks that
are summed over and finally to the rank of the left-hand side. This does
the right thing for products of two tensors. Use with products of more
than two tensors is experimental and currently untested.